refactor: added the users folder and transfered all user related controllers and their tests to it
